Refactoring: moved FeedHelpers.GetCatalogPropertiesAsync (#414)
