+++ tests/CatalogTests/Helpers/FeedHelpersTests.cs
@@ -24,157 +24,6 @@ namespace CatalogTests.Helpers
     {
         private const string _baseUri = "http://unit.test";
 
-        [Fact]
-        public async Task GetCatalogPropertiesAsync_ThrowsForNullStorage()
-        {
-            var exception = await Assert.ThrowsAsync<ArgumentNullException>(
-                () => FeedHelpers.GetCatalogPropertiesAsync(
-                    storage: null,
-                    telemetryService: Mock.Of<ITelemetryService>(),
-                    cancellationToken: CancellationToken.None));
-
-            Assert.Equal("storage", exception.ParamName);
-        }
-
-        [Fact]
-        public async Task GetCatalogPropertiesAsync_ThrowsForNullTelemetryService()
-        {
-            var exception = await Assert.ThrowsAsync<ArgumentNullException>(
-                () => FeedHelpers.GetCatalogPropertiesAsync(
-                    storage: Mock.Of<IStorage>(),
-                    telemetryService: null,
-                    cancellationToken: CancellationToken.None));
-
-            Assert.Equal("telemetryService", exception.ParamName);
-        }
-
-        [Fact]
-        public async Task GetCatalogPropertiesAsync_ThrowsIfCancelled()
-        {
-            await Assert.ThrowsAsync<OperationCanceledException>(
-                () => FeedHelpers.GetCatalogPropertiesAsync(
-                    Mock.Of<IStorage>(),
-                    Mock.Of<ITelemetryService>(),
-                    new CancellationToken(canceled: true)));
-        }
-
-        [Fact]
-        public async Task GetCatalogPropertiesAsync_ReturnsNullPropertiesIfStorageReturnsNull()
-        {
-            var storage = CreateStorageMock(json: null);
-
-            var catalogProperties = await FeedHelpers.GetCatalogPropertiesAsync(
-                storage.Object,
-                Mock.Of<ITelemetryService>(),
-                CancellationToken.None);
-
-            Assert.NotNull(catalogProperties);
-            Assert.Null(catalogProperties.LastCreated);
-            Assert.Null(catalogProperties.LastDeleted);
-            Assert.Null(catalogProperties.LastEdited);
-
-            storage.Verify();
-        }
-
-        [Fact]
-        public async Task GetCatalogPropertiesAsync_ReturnsNullPropertiesIfPropertiesMissing()
-        {
-            var storage = CreateStorageMock(json: "{}");
-
-            var catalogProperties = await FeedHelpers.GetCatalogPropertiesAsync(
-                storage.Object,
-                Mock.Of<ITelemetryService>(),
-                CancellationToken.None);
-
-            Assert.NotNull(catalogProperties);
-            Assert.Null(catalogProperties.LastCreated);
-            Assert.Null(catalogProperties.LastDeleted);
-            Assert.Null(catalogProperties.LastEdited);
-
-            storage.Verify();
-        }
-
-        [Fact]
-        public async Task GetCatalogPropertiesAsync_ReturnsAllPropertiesIfAllPropertiesSet()
-        {
-            var datetimeOffset = CreateDateTimeOffset(TimeSpan.Zero);
-            var json = $"{{\"nuget:lastCreated\":\"{(datetimeOffset.ToString("O"))}\"," +
-                $"\"nuget:lastDeleted\":\"{(datetimeOffset.ToString("O"))}\"," +
-                $"\"nuget:lastEdited\":\"{(datetimeOffset.ToString("O"))}\"}}";
-            var storage = CreateStorageMock(json);
-
-            var catalogProperties = await FeedHelpers.GetCatalogPropertiesAsync(
-                storage.Object,
-                Mock.Of<ITelemetryService>(),
-                CancellationToken.None);
-
-            Assert.NotNull(catalogProperties);
-            Assert.NotNull(catalogProperties.LastCreated);
-            Assert.NotNull(catalogProperties.LastDeleted);
-            Assert.NotNull(catalogProperties.LastEdited);
-            Assert.Equal(datetimeOffset, catalogProperties.LastCreated.Value);
-            Assert.Equal(datetimeOffset, catalogProperties.LastDeleted.Value);
-            Assert.Equal(datetimeOffset, catalogProperties.LastEdited.Value);
-
-            storage.Verify();
-        }
-
-        [Fact]
-        public async Task GetCatalogPropertiesAsync_ReturnsDateTimeWithFractionalHourOffsetInUtc()
-        {
-            var lastCreated = CreateDateTimeOffset(new TimeSpan(hours: 5, minutes: 30, seconds: 0));
-            var json = CreateCatalogJson("nuget:lastCreated", lastCreated);
-
-            await VerifyPropertyAsync(json, catalogProperties =>
-                {
-                    Assert.NotNull(catalogProperties.LastCreated);
-                    Assert.Equal(lastCreated, catalogProperties.LastCreated.Value);
-                    Assert.Equal(DateTimeKind.Utc, catalogProperties.LastCreated.Value.Kind);
-                });
-        }
-
-        [Fact]
-        public async Task GetCatalogPropertiesAsync_ReturnsDateTimeWithPositiveOffsetInUtc()
-        {
-            var lastDeleted = CreateDateTimeOffset(TimeSpan.FromHours(1));
-            var json = CreateCatalogJson("nuget:lastDeleted", lastDeleted);
-
-            await VerifyPropertyAsync(json, catalogProperties =>
-                {
-                    Assert.NotNull(catalogProperties.LastDeleted);
-                    Assert.Equal(lastDeleted, catalogProperties.LastDeleted.Value);
-                    Assert.Equal(DateTimeKind.Utc, catalogProperties.LastDeleted.Value.Kind);
-                });
-        }
-
-        [Fact]
-        public async Task GetCatalogPropertiesAsync_ReturnsDateTimeWithNegativeOffsetInUtc()
-        {
-            var lastEdited = CreateDateTimeOffset(TimeSpan.FromHours(-1));
-            var json = CreateCatalogJson("nuget:lastEdited", lastEdited);
-
-            await VerifyPropertyAsync(json, catalogProperties =>
-                {
-                    Assert.NotNull(catalogProperties.LastEdited);
-                    Assert.Equal(lastEdited, catalogProperties.LastEdited.Value);
-                    Assert.Equal(DateTimeKind.Utc, catalogProperties.LastEdited.Value.Kind);
-                });
-        }
-
-        [Fact]
-        public async Task GetCatalogPropertiesAsync_ReturnUtcDateTimeInUtc()
-        {
-            var lastCreated = DateTimeOffset.UtcNow;
-            var json = CreateCatalogJson("nuget:lastCreated", lastCreated);
-
-            await VerifyPropertyAsync(json, catalogProperties =>
-                {
-                    Assert.NotNull(catalogProperties.LastCreated);
-                    Assert.Equal(lastCreated, catalogProperties.LastCreated.Value);
-                    Assert.Equal(DateTimeKind.Utc, catalogProperties.LastCreated.Value.Kind);
-                });
-        }
-
         public static IEnumerable<object[]> GetPackages_GetsAllPackages_data
         {
             get
@@ -757,46 +606,6 @@ namespace CatalogTests.Helpers
                     }
                 };
             }
-        }
-
-        private static string CreateCatalogJson(string propertyName, DateTimeOffset propertyValue)
-        {
-            return $"{{\"{propertyName}\":\"{(propertyValue.ToString("O"))}\"}}";
-        }
-
-        private static DateTimeOffset CreateDateTimeOffset(TimeSpan offset)
-        {
-            var datetime = new DateTime(DateTime.Now.Ticks, DateTimeKind.Unspecified);
-
-            return new DateTimeOffset(datetime, offset);
-        }
-
-        private static Mock<IStorage> CreateStorageMock(string json)
-        {
-            var storage = new Mock<IStorage>(MockBehavior.Strict);
-
-            storage.Setup(x => x.ResolveUri(It.IsNotNull<string>()))
-                .Returns(new Uri("https://unit.test"))
-                .Verifiable();
-            storage.Setup(x => x.LoadStringAsync(It.IsNotNull<Uri>(), It.IsAny<CancellationToken>()))
-                .ReturnsAsync(json)
-                .Verifiable();
-
-            return storage;
-        }
-
-        private static async Task VerifyPropertyAsync(string json, Action<CatalogProperties> propertyVerifier)
-        {
-            var storage = CreateStorageMock(json);
-
-            var catalogProperties = await FeedHelpers.GetCatalogPropertiesAsync(
-                storage.Object,
-                Mock.Of<ITelemetryService>(),
-                CancellationToken.None);
-
-            Assert.NotNull(catalogProperties);
-            propertyVerifier(catalogProperties);
-            storage.Verify();
         }
 
         private static NupkgMetadata GetNupkgMetadata(string resourceName)
